feat(deploy): add checking of opening ports Changes are as follows: 	add functions isLocalPortOpen and isRemotePortOpen 	add using this functions in code 	rewrite log messages for checking ports 	replace struct sshTunnel and portForward with tunnelToCluster 	add stdout in terraformDestroy 	add defer function closePortForward()
Checking ports 8080 (for Grafana) and 6443(for k8s) is nessesary
to avoid inconsistent state of cluster deployment, when the cluster
is successfully deployed, but there is no access to it from the local machine.
If ports 8080 and 6443 are busy, we use reverse ports - 8081 and 6444.
If this ports are busy also then deploy will be failed with log message.