feat(ssh tunnel): implement ssh tunneling instead of using os.Exec(<ssh cmd>)
`os.Exec(<ssh tunnel cmd>)` was used for PoC to speed up developing process.
The main problem we face with this is the inability to open ssh tunnel
synchronously (we have to keep tunnel in the background, so we used
`os.Start` which is async).

Current ssh tunnel implementation is based on
https://github.com/elliotchance/sshtunnel with the following editions:

1) Changed SSHTunnel.Start() signature to SSHTunnel.Start(chan err)
    That is the way to wait for tunnel to be started or to fail to do so.
2) Proxy connection over single ssh connection instead of establishing
    new SSH connection for every TCP connection.